Add LogController; secure Docker & update frontend
Add a new LogController (api/logs/{limit}) to return authenticated subscriber notification logs and remove the duplicate logs action from TemplatesController. Introduce docker-compose.override.yml to expose DB/Redis/RabbitMQ ports via env vars and harden docker-compose.yml by adding a security-strict anchor, applying security_opt (no-new-privileges), dropping capabilities, and adding NET_BIND_SERVICE to the frontend container; port mappings were moved to the override file. Update frontend Dockerfile to remove the nginx user line, fix ownership of nginx runtime directories, touch pid file and switch to the nginx user for safer runtime. Also update Plan.drawio diagram metadata/layout.
