+++ docker-compose.yml
@@ -1,3 +1,9 @@
+x-security-strict: &security-strict
+  security_opt:
+    - no-new-privileges:true
+  cap_drop:
+    - ALL
+
 services:
   frontend:
     container_name: frontend
@@ -22,6 +28,12 @@ services:
       options:
         max-size: "10m"
         max-file: "3"
+    security_opt:
+      - no-new-privileges:true
+    cap_drop:
+      - ALL
+    cap_add:
+      - NET_BIND_SERVICE
 
   api:
     container_name: api
@@ -68,13 +80,12 @@ services:
           memory: 400M
     logging:
       <<: *default-logging
+    <<: *security-strict
 
   redis:
     container_name: redis
     image: redis:7-alpine
     command: redis-server --maxmemory 50mb --maxmemory-policy allkeys-lru
-    ports:
-      - "${REDIS_PORT:-6379}:6379"
     volumes:
       - redis_data:/data
     networks:
@@ -91,13 +102,12 @@ services:
           memory: 80M
     logging:
       <<: *default-logging
+    security_opt:
+      - no-new-privileges:true
 
   rabbitmq:
     image: rabbitmq:3-management-alpine
     container_name: rabbitmq
-    ports:
-      - "5672:5672"
-      - "15672:15672"
     environment:
       - RABBITMQ_DEFAULT_USER=${RABBITMQ_USER:-guest}
       - RABBITMQ_DEFAULT_PASS=${RABBITMQ_PASS:-guest}
@@ -117,6 +127,8 @@ services:
           memory: 300M
     logging:
       <<: *default-logging
+    security_opt:
+      - no-new-privileges:true
 
   db:
     container_name: db
@@ -127,8 +139,6 @@ services:
       - POSTGRES_DB=${DB_NAME}
     command: >
       postgres -c shared_buffers=64MB -c effective_cache_size=128MB -c work_mem=4MB -c maintenance_work_mem=32MB
-    ports:
-      - "${DB_PORT:-5432}:5432"
     volumes:
       - postgres_data:/var/lib/postgresql/data
     networks:
@@ -145,6 +155,8 @@ services:
           memory: 300M
     logging:
       <<: *default-logging
+    security_opt:
+      - no-new-privileges:true
 
   tunnel:
     image: cloudflare/cloudflared:latest
@@ -162,6 +174,7 @@ services:
           memory: 64M
     logging:
       <<: *default-logging
+    <<: *security-strict
 
 networks:
   fertile-network:
