+++ frontend/Dockerfile
@@ -20,5 +20,16 @@ RUN echo 'server { \
     } \
     }' > /etc/nginx/conf.d/default.conf
 
+RUN sed -i '/user  nginx;/d' /etc/nginx/nginx.conf
+
+RUN chown -R nginx:nginx /usr/share/nginx/html && \
+    chown -R nginx:nginx /var/cache/nginx && \
+    chown -R nginx:nginx /var/log/nginx && \
+    chown -R nginx:nginx /etc/nginx/conf.d && \
+    touch /var/run/nginx.pid && \
+    chown -R nginx:nginx /var/run/nginx.pid
+
+USER nginx
+
 EXPOSE 80
 CMD ["nginx", "-g", "daemon off;"]
